Created entrypoint.sh script.
`entrypoint.sh` is the file that will be used by the container to start the Archivy process. If the argument "start" is passed to the argument, the script will start Archivy. If any other arguments are passed, they will be executed instead of Archivy. Please read the preamble in the `entrypoint.sh` file for more information.