Added documentation for Docker. Other changes too
Added a `.dockerignore` file which is useful when building images
locally. The files in `.dockerignore` will not be sent to the Docker
daemon during build time, thereby making the final image smaller.

Added `Docker.md` which is going to be the source of all documentation
for Archivy pertaining to containers.

Have added a few lines in `Dockerfile` to get rid of some of the large,
unnecessary repository files like `archivy.gif` and other documentation
files which are not needed for running Archivy inside the container.

Have removed the title comments from the compose files. This causes
erros with 'Play With Docker'.

Have added a compose file `docker-compose-with-elasticsearch.yaml` which
takes care of setting up Archivy with Elasticsearch.