Fixed issues with entrypoint and compose files
The `entrypoint.sh` file did not have the execute bit set on it due to
which it could not be executed within the container. This caused the
container startup to fail. Have set the execute bit on `entrypoint.sh`
file now.

Have fixed some errors in `docker-compose-with-elasticsearch.yaml` file.
Changed `attemps` to `attempts`, changed the `ports` value to a list.

Have fixed the `docker-compose` commands in the documentation. The order
of the arguments was incorrect previously. To specify a compose file,
the `-f` flag must be provided before the `up`, `down`, or any other
commands.

Changed
    `docker-compose up -d -f ./docker-compose.yaml`
to
    `docker-compose -f ./docker-compose.yaml up -d`