Major additions to entrypoint.sh
Since Archivy was not able to connect to Elasticsearch due to it taking
some time to start, a health check of sorts has been added to the
`entrypoint.sh` file. There are now two functions in the script which
will check if Elasticsearch is up and running at the URL indicated in
the ELASTICSEARCH_URL variable. Until Elasticsearch responds to
requests, the script will not start Archivy. This is done *only if* the
ELASTICSEARCH_ENABLED variable is set to `1`. That way, the starting of
Archivy will not be blocked if Elasticsearch support is not required.
The health check function also requires the existence of either
`netcat` or `curl` in the container. It is flexible enough to use either
one, depending on the whichever it present in the container.

Have made changes to both `Dockerfile` and `local-build.Dockerfile` to
add tools necessary to perform the aforementioned health checks on
Elasticsearch. In `local-build.Dockerfile`, installation of `netcat` has
been added. In `Dockerfile`, installation of `curl` has been added.

Have added `docker-compose*` to `.dockerignore` to ensure no Docker
Compose files are transferred to the Docker daemon during build time.