Remove redundant / branched requirements.txt
The requirements.txt required some modules which was not directly used within
archivy. Some requirements are the direct dependencies of the dependencies.
Pinned versions on requirements sometimes creates conflicts on end-users's pip
installation. For example, importlib_metadata==1.6.0 has a buggy pkg_resources.DistributionNotFound , which is now fixed in importlib_metadata==1.7.0