Version 0.0.7:
Small version because of this urgent bug:

When the `ELASTICSEARCH_ENABLED` option was set, even if set to 0, it
would be saved as a string which are always truthy in python. This meant
that setting the value to 0 would still run elasticsearch.

Other feature:

When creating a new note or bookmark, it will now open your editor with
the md file.