Major changes to Dockerfile and entrypoint.sh
The new Docker image is 2.5 times smaller now because the new Dockerfile
uses multi-stage builds. Now, all necessary binaries and libraries are
installed in the first stage, and the required files are copied over to
a second stage which is built from a much lighter Alpine Linux image.
Also, the process no longer runs as `root` inside the container. Have
created a non-root user specifically for running Archivy.

`entrypoint.sh` has been modified to be compatible with `sh`. It's
functionality is more or less the same as before.

Have edited compose file to use the new Docker image.