Improved Dockerfile.
Have modified the Dockerfile so that the path to the directory in which
data will be saved by Archivy is easier to type when compared to before.
Previously it was `~/.local/share/archivy/data` which is a bit long and
annoying to type as part of the `docker run` command. Now, the path to
the data directory is `/archivy/data` which is much simpler to remember
and type. Also, the binaries and libraries from the builder stage are
moved to `/usr/local` which is already in PATH. Therefore, the step
involving the setting of PATH in Dockerfile has been removed.

Have made changes to `entrypoint.sh` to help change the data directory
in Archivy by setting the variable `ARCHIVY_DATA_DIR` to `/archivy`.

Have made required changes to `container-structure-test.yaml`.