Added healthcheck to Dockerfile. Bumped up version.
Added the `HEALTHCHECK` directive to `Dockerfile`. To that end, I've
also added the `healthcheck.sh` which is used by the Docker daemon to
check if the container is healthy i.e up and running.

Have bumped up the version of Archivy in the container to `0.1.0`. These
changes have been made to `container-structure-test.yaml`,
`docker-compose.yml`, `docker-compose-with-elasticsearch.yml`, and
`Dockerfile`.