Dockerfile for `master` created. Other changes too
A new Dockerfile titled `Dockerfile.source` has been added. This
Dockerfile is used to build the image from the `master` branch.

Made some improvements to `entrypoint.sh` and `healthcheck.sh`. Minor
additions to `container-structure-test.yaml`.

`Docker.md` has been added which serves as the documentation for the
image on DockerHub.

Two Github Actions files have been added. `build-on-release.yml` is used
to build, test, and push images with the appropriate versioned tags
whenever a new release is published. `build-on-push-master.yml` is used
to build, test, and push images with the `source` tag whenever a commit
is made to the code in the `master` branch. This image will be built
directly from the source code in the `master` branch.