Fixed errors in actions files. New file added.
Have fixed errors in the `build-on-release.yml` and
`build-on-push-master.yml` files. The `checkout` step was checking out
the `master` branch of the repository, when in fact, it should have been
checking out the `docker` branch. Have added the necessary options to
ensure that the `docker` branch is checkout out.

Have removed the checkout step from the `deleteUntestedImage` job as it
isn't necessary.

Have added `update-dockerhub-description.yml` which is triggered
whenever a push event occurs on the `docker` branch AND if the
`Docker.md` file is changed. This file updates the description of the
Archivy image on Docker Hub. This way, the description on Docker Hub can
be in sync with the `Docker.md` file.