Fixed 'update...description.sh' script
Fixed major errors in the `update-requirements-description.sh` script.
This script is used to replace placeholder text in a file with the
content obtained from the `requirements.txt` file in the Archivy
repository on Github.

All comments and empty lines are removed and it is formatted as bullet
points in Markdown.