Dockerfile for `master` created. Other changes too (#87)
* Dockerfile for `master` created. Other changes too

A new Dockerfile titled `Dockerfile.source` has been added. This
Dockerfile is used to build the image from the `master` branch.

Made some improvements to `entrypoint.sh` and `healthcheck.sh`. Minor
additions to `container-structure-test.yaml`.

`Docker.md` has been added which serves as the documentation for the
image on DockerHub.

Two Github Actions files have been added. `build-on-release.yml` is used
to build, test, and push images with the appropriate versioned tags
whenever a new release is published. `build-on-push-master.yml` is used
to build, test, and push images with the `source` tag whenever a commit
is made to the code in the `master` branch. This image will be built
directly from the source code in the `master` branch.

* Fixed errors in actions files. New file added.

Have fixed errors in the `build-on-release.yml` and
`build-on-push-master.yml` files. The `checkout` step was checking out
the `master` branch of the repository, when in fact, it should have been
checking out the `docker` branch. Have added the necessary options to
ensure that the `docker` branch is checkout out.

Have removed the checkout step from the `deleteUntestedImage` job as it
isn't necessary.

Have added `update-dockerhub-description.yml` which is triggered
whenever a push event occurs on the `docker` branch AND if the
`Docker.md` file is changed. This file updates the description of the
Archivy image on Docker Hub. This way, the description on Docker Hub can
be in sync with the `Docker.md` file.

* Minor changes and error fixes.

* Added file to list issues faced in building images.

* Removed multi-arch support

Have removed support for multiple architectures as `pandoc` is not
available on Alpine Linux for architectures other than `amd64`.
Therefore, for versions>=`0.2.0`, images only for `amd64` are available.

In the future, I will try to compile `pandoc` for the other
architectures and check if the image works.

* Fixed 'update...description.sh' script

Fixed major errors in the `update-requirements-description.sh` script.
This script is used to replace placeholder text in a file with the
content obtained from the `requirements.txt` file in the Archivy
repository on Github.

All comments and empty lines are removed and it is formatted as bullet
points in Markdown.

* Major improvements to 'updare-dockerhub-readme.sh' script. Other changes in Actions files too.

* Have set ARCHIVY_PORT=5000 in entrypoint.sh.