Changes to docker-related Actions files
Have removed the "on PR to master branch" condition in the
`build-docker-on-push-master.yml` file as this was causing the builds to
run when a PR was created. This could result in images being pushed to
DockerHub with changes that were not yet merged into `master`.

Have changed the version of the `checkout` action from `v1` to `v2` in
`build-docker-on-push-master.yml` and `build-docker-on-release.yml`.