Fixed errors in actions files.
In `build-docker-on-push-master.yml`, the job `buildPushUntested` was
referring to a variable `tag_name` instead of `image_tag`. Fixed this.

In `build-docker-on-release.yml`, the conditions which trigger the run
were `on publish of release`. This wasn't getting triggered when tags
were being pushed, which is needed. Have added a condition `on push of
tags`.

Also, there's an error in the linting step(`Hadolint`). This will be
fixed in the `docker` branch.