Fixed errors in Actions files (#118)
* Fixed errors in both actions.yml files. There was a typo in the job dependency name.

* Changes to docker-related Actions files

Have removed the "on PR to master branch" condition in the
`build-docker-on-push-master.yml` file as this was causing the builds to
run when a PR was created. This could result in images being pushed to
DockerHub with changes that were not yet merged into `master`.

Have changed the version of the `checkout` action from `v1` to `v2` in
`build-docker-on-push-master.yml` and `build-docker-on-release.yml`.

* Fixed errors in docker-related actions files

In files `build-docker-on-push-master.yml` and
`build-docker-on-release.yml`, the `--build-arg` value needs to be in
the form of `KEY=VALUE`. Instead, it was `KEY VALUE`. Have fixed this.

* Migrated to official actions where possible in actions files.

* Changed Dockerfile to Dockerfile.source in action file which builds image based on master branch.

* Fixed errors in actions files.

In `build-docker-on-push-master.yml`, the job `buildPushUntested` was
referring to a variable `tag_name` instead of `image_tag`. Fixed this.

In `build-docker-on-release.yml`, the conditions which trigger the run
were `on publish of release`. This wasn't getting triggered when tags
were being pushed, which is needed. Have added a condition `on push of
tags`.

Also, there's an error in the linting step(`Hadolint`). This will be
fixed in the `docker` branch.