Fixed errors in docker Actions files (#136)
In `build-docker-on-release.yml`, the trigger condition has been changed
to run _only_ when a release is published. Previously, the build would
get triggered when a tag was also pushed. This resulted in the build
step running before a release was created on Github.
The `pypi-auto-deploy.yml` file handles the creation and publishing of
the release when a new tag is pushed. Therefore, it is sufficient to
trigger is build of the docker image only when a release is published.

In both `build-docker-on-release.yml` and
`build-docker-on-push-master.yml`, the environment variable setting step
has been changed as `set-env` has been deprecated and disabled. Now, I
write the required data to files using `echo "data" > VARIABLE`, and
then access the data by using `cat VARIABLE`.