Updated Python version. Other changes too
The version of python in the base image has been upgraded from `3.8` to
`3.9` in both `Dockerfile` and `Dockerfile.source`.
Some changes have been made to `Dockerfile` and to `entrypoint.sh in
order to allow the user to use the `archivy` CLI.
`musl` has been upgraded to the next version due to a CVE in the default
version that comes with the base image of `python3.9-alpine`.
`container-structure-test.yaml` has been updated to reflect the changes
in the Dockerfiles.
An issue has been added to `issues.md`.