Error fix in actions files
In `build-docker-on-push-master.yml` and `build-docker-on-release.yml`,
an extra step has been added in the 'Test and Scan' job. This step
outputs the name of the image to an environment variable so that it can
be accessed by `container-structure-test` and others in that job.

This is being done as command substitution does not work in the `image`
value of `container-stucture-test`. Hopefully, variable expansion works.