fix: 'archivy init' does not purge old configuration before interactive init
archivy does not reset the configuration before creating the new user.
This is because archivy calls click.create_admin before resetting the configuration.
It is also recommended to purge the configuration, (and not replace existing conf), to
remove old entries

along with this patch, a new value, VERSION would be added to the configuration
which would be useful in future to migrate old configuration to new configuration tree.