v1.0
Archivy began as a small personal project in August 2020. However, since then it's grown into much more, with the help of contributors and the many discussions and ideas we've brainstormed on the discord server (https://discord.gg/uQsqyxB).

There have been major improvements in the design, structure and general features
of the project like for example:

- development of a CLI interface and deployment of archivy on pip, docker, the aur and nixpkgs
- redesign to make archivy actually look good :slight_smile:
- introduction of a powerful markdown editor / rendering
- a complete documentation website
- auth system
- focus on extensibility: Python API / Web API / plugin system

and much more!

I'm extremely excited for the current state of Archivy and it's future
even beyond v1, outlined in the roadmap https://github.com/archivy/archivy/issues/74#issuecomment-764828063.

Thank you to all the people who have shaped this project, and as
always, don't hesitate to ask any questions or talk to us if you'd like to share your opinion or ideas!